Refactors SettingsSection to SettingsScreen; Moves navigate calls from SettingsModal to SettingsScreen; Removes SettingsModal.
