+++ src/components/settings-menu/index.js
@@ -2,4 +2,3 @@ export { default as CurrencySection } from './CurrencySection';
 export { default as NetworkSection } from './NetworkSection/NetworkSection';
 export { default as NotificationsSection } from './NotificationsSection';
 export { default as WalletConnectSessionsSection } from './WalletConnectSessionsSection';
-export { default as SettingsSection } from './SettingsSection';
